Execute `afterAll` despite pending examples.
Example groups would only execute `afterAll` blocks when the number of run
examples was equal to the total number of examples. Pending examples are
never run, so this condition was never met for example groups containing
pending examples.

Add `pendingExampleCount` and change condition to execute `afterAll`
blocks when `(ran + pending) = total`. Execute `afterAll` blocks whether
the example group contains pending examples or not.

Also add a test which fails without the new conditional, but passes when
the new conditional is used.

Note that `beforeAll` and `afterAll` blocks are *not* executed when all
examples in the group are pending. This is not consistent with RSpec
behavior, which executes `before(:all)` and `after(:all)` blocks even if
all examples in the group are pending (as of 3.0.0).

Addresses https://github.com/specta/specta/issues/61.