Specta.m: raise when shared example does not exist
Shared examples do not fail when the specified shared example group is
not defined. The original, intended behavior in case a user tried to use
an undefined shared example group was to raise an exception. [1] This
behavior was broken during the migration from SenTestingKit to XCTest. [2]
The regression was caused by a failure to "re-evaluate"
SPTCurrentTestCase, which is nil in some cases.

- Re-evaluate SPTCurrentTestCase within an it block in order to inform
  the XCTest harness of a test failure.
- Add SharedExamplesTest6 to prevent future regressions.

Fixes issue #68.

[1] https://github.com/specta/specta/commit/db0ec3d7#diff-5f0b31e53d8721b15d8afed4265ce1caR138
[2] https://github.com/specta/specta/commit/725ba21b#diff-5f0b31e53d8721b15d8afed4265ce1caR136