Whitelist classes for global before/after hooks
When running global before and after hooks, we were picking up any class
that implented +beforeEach and/or +afterEach. You could opt-out by
conforming to the SPTExcludeGlobalBeforeAfterEach protocol. This could
cause difficult to debug failures when it picked up imported classes.

This commit reverses the approach, creating a
SPTIncludeGlobalBeforeAfterEach protocol, which SPTExampleGroup
implements. Now we only select classes that implement
SPTIncludeGlobalBeforeAfterEach for running global +beforeEach and/or
+afterEach. You can opt-in by implementing the protocol.